+++ src/backend/xrender/xrender.c
@@ -956,8 +956,9 @@ static backend_t *xrender_init(session_t *ps, xcb_window_t target) {
 	xd->back_image.pict = xd->vsync ? xd->back[xd->curr_back] : xd->target;
 
 	auto drivers = detect_driver(xd->base.c->c, &xd->base, xd->target_win);
-	if (drivers & DRIVER_MODESETTING) {
-		// I believe other xf86-video drivers have accelerated blur?
+	if (drivers & (DRIVER_MODESETTING | DRIVER_NVIDIA | DRIVER_NOUVEAU |
+	               DRIVER_AMDGPU | DRIVER_RADEON | DRIVER_FGLRX)) {
+		// I believe xf86-video-intel have accelerated convolution?
 		xd->quirks |= BACKEND_QUIRK_SLOW_BLUR;
 	}